Software service provider
This website takes a website and loads the content on the website and stores it in the database. Based on that, a chat bot will ask the users different questions to know about their product and ask if they want to use their developers to do a task. This website is like a bridge between a client and a software developer.